Move `ShardHashRing` (and related items) from `shard_holder.rs` into `hash_ring.rs`
