+++ lib/collection/src/shards/shard_holder.rs
@@ -1,6 +1,5 @@
 use std::collections::{HashMap, HashSet};
 use std::fmt;
-use std::hash::Hash;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
@@ -8,7 +7,6 @@ use common::cpu::CpuBudget;
 use itertools::Itertools;
 // TODO rename ReplicaShard to ReplicaSetShard
 use segment::types::ShardKey;
-use smallvec::SmallVec;
 use tar::Builder as TarBuilder;
 use tokio::runtime::Handle;
 use tokio::sync::RwLock;
@@ -17,7 +15,7 @@ use super::replica_set::AbortShardTransfer;
 use super::transfer::transfer_tasks_pool::TransferTasksPool;
 use crate::common::validate_snapshot_archive::validate_open_snapshot_archive;
 use crate::config::{CollectionConfig, ShardingMethod};
-use crate::hash_ring::HashRing;
+use crate::hash_ring::ShardHashRing;
 use crate::operations::shard_selector_internal::ShardSelectorInternal;
 use crate::operations::shared_storage_config::SharedStorageConfig;
 use crate::operations::snapshot_ops::SnapshotDescription;
@@ -33,18 +31,10 @@ use crate::shards::shard_versioning::latest_shard_paths;
 use crate::shards::transfer::{ShardTransfer, ShardTransferKey};
 use crate::shards::CollectionId;
 
-const HASH_RING_SHARD_SCALE: u32 = 100;
-
 const SHARD_TRANSFERS_FILE: &str = "shard_transfers";
 pub const SHARD_KEY_MAPPING_FILE: &str = "shard_key_mapping.json";
 
 pub type ShardKeyMapping = HashMap<ShardKey, HashSet<ShardId>>;
-
-/// List type for shard IDs
-///
-/// Uses a `SmallVec` putting two IDs on the stack. That's the maximum number of shards we expect
-/// with the current resharding implementation.
-pub type ShardIds = SmallVec<[ShardId; 2]>;
 
 pub struct ShardHolder {
     shards: HashMap<ShardId, ShardReplicaSet>,
@@ -55,95 +45,6 @@ pub struct ShardHolder {
     // Duplicates the information from `key_mapping` for faster access
     // Do not require locking
     shard_id_to_key_mapping: HashMap<ShardId, ShardKey>,
-}
-
-#[derive(Clone)]
-pub enum ShardHashRing {
-    /// Single hashring
-    Single(HashRing<ShardId>),
-
-    /// Two hashrings when transitioning during resharding
-    /// Depending on the current resharding state, points may be in either or both shards.
-    Resharding {
-        old: HashRing<ShardId>,
-        new: HashRing<ShardId>,
-    },
-}
-
-impl ShardHashRing {
-    /// Create a new single hashring.
-    ///
-    /// The hashring is created with a fair distribution of points and `HASH_RING_SHARD_SCALE` scale.
-    pub fn single() -> Self {
-        Self::Single(HashRing::fair(HASH_RING_SHARD_SCALE))
-    }
-
-    /// Create a new resharding hashring, with resharding shard already added into `new` hashring.
-    ///
-    /// The hashring is created with a fair distribution of points and `HASH_RING_SHARD_SCALE` scale.
-    pub fn resharding(shard_id: ShardId) -> Self {
-        let mut ring = Self::Resharding {
-            old: HashRing::fair(HASH_RING_SHARD_SCALE),
-            new: HashRing::fair(HASH_RING_SHARD_SCALE),
-        };
-
-        ring.add_resharding(shard_id);
-
-        ring
-    }
-
-    pub fn is_empty(&self) -> bool {
-        match self {
-            Self::Single(ring) => ring.is_empty(),
-            Self::Resharding { old, new } => old.is_empty() && new.is_empty(),
-        }
-    }
-
-    pub fn is_resharding(&self) -> bool {
-        matches!(self, Self::Resharding { .. })
-    }
-
-    pub fn add(&mut self, shard: ShardId) {
-        match self {
-            Self::Single(ring) => ring.add(shard),
-            Self::Resharding { old, new } => {
-                if new.get(&shard).is_none() {
-                    old.add(shard);
-                    new.add(shard);
-                }
-            }
-        }
-    }
-
-    pub fn add_resharding(&mut self, shard: ShardId) {
-        if let Self::Single(ring) = self {
-            let (old, new) = (ring.clone(), ring.clone());
-            *self = Self::Resharding { old, new };
-        }
-
-        let Self::Resharding { new, .. } = self else {
-            unreachable!();
-        };
-
-        new.add(shard);
-    }
-
-    pub fn get<U: Hash>(&self, key: &U) -> ShardIds {
-        match self {
-            Self::Single(ring) => ring.get(key).into_iter().cloned().collect(),
-            // TODO(resharding): just use the old hash ring for now, never route to two shards
-            // TODO(resharding): switch to both as commented below once read folding is implemented
-            Self::Resharding { old, new: _ } => old.get(key).into_iter().cloned().collect(),
-            // Self::Resharding { old, new } => old
-            //     .get(key)
-            //     .into_iter()
-            //     .chain(new.get(key))
-            //     // Both hash rings may return the same shard ID, take it once
-            //     .dedup()
-            //     .cloned()
-            //     .collect(),
-        }
-    }
 }
 
 pub type LockedShardHolder = RwLock<ShardHolder>;
